Cleaned up main Create component
I moved the useEffect and picker to separate files and created a folder for the Create component.
