+++ WarPortal/client/src/App.jsx
@@ -7,7 +7,7 @@ import { StoryArcs } from "./components/Story.jsx";
 import { ContactUs } from "./components/ContactUs.jsx";
 import { Admin } from "./components/Admin.jsx";
 import { SignIn } from "./components/SignIn.jsx";
-import { Create } from "./components/Create.jsx";
+import { Create } from "./components/Create/Create.jsx";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Footer from "./components/Footer.jsx";
 import CardValidator from "./components/CardValidator";
